LPS-76475 Move util class to internal package and avoid major bump
