+++ modules/util/css-builder/src/main/java/com/liferay/css/builder/CSSBuilderUtil.java
@@ -15,78 +15,40 @@
 package com.liferay.css.builder;
 
 import com.liferay.css.builder.internal.util.StringUtil;
-import com.liferay.petra.string.CharPool;
-import com.liferay.petra.string.StringBundler;
-import com.liferay.petra.string.StringPool;
-
-import java.io.File;
-
-import java.util.Date;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 /**
  * @author Brian Wing Shun Chan
  * @author Raymond Augé
  * @author Eduardo Lundgren
  * @author Shuyang Zhou
+ * @deprecated As of 2.1.0, replaced by
+ *                          #{@link com.liferay.css.builder.internal.util.CSSBuilderUtil}
  */
-public class CSSBuilderUtil {
+@Deprecated
+public class CSSBuilderUtil
+	extends com.liferay.css.builder.internal.util.CSSBuilderUtil {
 
-	public static File getOutputFile(String fileName, String outputDirName) {
-		return getOutputFile(fileName, outputDirName, StringPool.BLANK);
+	public static String parseStaticTokens(String content) {
+		return StringUtil.replace(
+			content,
+			new String[] {
+				"@model_hints_constants_text_display_height@",
+				"@model_hints_constants_text_display_width@",
+				"@model_hints_constants_textarea_display_height@",
+				"@model_hints_constants_textarea_display_width@"
+			},
+			new String[] {
+				_TEXT_DISPLAY_HEIGHT, _TEXT_DISPLAY_WIDTH,
+				_TEXTAREA_DISPLAY_HEIGHT, _TEXTAREA_DISPLAY_WIDTH
+			});
 	}
 
-	public static File getOutputFile(
-		String fileName, String outputDirName, String suffix) {
+	private static final String _TEXT_DISPLAY_HEIGHT = "15";
 
-		return new File(getOutputFileName(fileName, outputDirName, suffix));
-	}
+	private static final String _TEXT_DISPLAY_WIDTH = "210";
 
-	public static String getOutputFileName(
-		String fileName, String outputDirName, String suffix) {
+	private static final String _TEXTAREA_DISPLAY_HEIGHT = "100";
 
-		String cacheFileName = StringUtil.replace(
-			fileName, CharPool.BACK_SLASH, CharPool.SLASH);
-
-		int x = cacheFileName.lastIndexOf(CharPool.SLASH);
-		int y = cacheFileName.lastIndexOf(CharPool.PERIOD);
-
-		if (cacheFileName.endsWith(".scss")) {
-			cacheFileName = cacheFileName.substring(0, y + 1) + "css";
-		}
-
-		return StringBundler.concat(
-			cacheFileName.substring(0, x + 1), outputDirName,
-			cacheFileName.substring(x + 1, y), suffix,
-			cacheFileName.substring(y));
-	}
-
-	public static String getRtlCustomFileName(String fileName) {
-		int pos = fileName.lastIndexOf(CharPool.PERIOD);
-
-		return fileName.substring(0, pos) + "_rtl" + fileName.substring(pos);
-	}
-
-	public static String parseCSSImports(String content) {
-		StringBuffer sb = new StringBuffer();
-
-		Matcher matcher = _cssImportPattern.matcher(content);
-
-		Date date = new Date();
-
-		while (matcher.find()) {
-			String cssImport = matcher.group();
-
-			matcher.appendReplacement(sb, cssImport + "?t=" + date.getTime());
-		}
-
-		matcher.appendTail(sb);
-
-		return sb.toString();
-	}
-
-	private static final Pattern _cssImportPattern = Pattern.compile(
-		"@import\\s+url\\s*\\(\\s*['\"]?(.+\\.css)");
+	private static final String _TEXTAREA_DISPLAY_WIDTH = "500";
 
 }
